src/collectors/requirements.txt
beautifulsoup4==4.8.1
bleach==4.1.0
certifi==2021.10.8
feedparser==5.2.1
Flask==1.1.1
Flask-Cors==3.0.10
Flask-RESTful==0.3.7
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.0.4
lxml==4.6.5
marshmallow==3.5.0
marshmallow-enum==1.5.1
pyslack==0.5.0
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.5
pytz==2019.3
requests==2.26.0
schedule==0.6.0
selenium==4.0.0
six==1.14.0
slackclient==1.0.7
soupsieve==1.9.5
tweepy==3.8.0
Werkzeug==0.16.0
zipp==3.1.0
dateparser==1.1.0
